requirements/local.txt
-r base.txt

ipython>=8.0
pytest>=8.0
pytest-django>=4.8
